requirements.txt
absl-py
datasets
drawdata
editdistance
einops
matplotlib
numpy
pyyaml
schedulefree
scikit-learn
scipy
simplipy
torch
torch_optimizer
tqdm
wandb
zss
